requirements.txt
Flask
geopy
gspread
oauth2client
pandas
# scikit-learn
scikit-learn==0.23.2
gunicorn
# gunicorn==20.0.4
requests
keras
tensorflow
numpy
pillow

# Jinja2==2.8
# MarkupSafe==0.23
# Werkzeug==0.11.11
# click==6.6
# itsdangerous==0.24
# requests
# numpy
# clint
# scipy
# flask_restful
# flask-jsonpify
# beautifulsoup4
# rq
# timeout-decorator
# pythainlp
# xlrd
# botnoi==0.2.9
# matplotlib
# gensim